chore: merged files of conn and dax
